+++ fe/src/components/Header/SearchBar/MiniSearchBar/index.tsx
