+++ fe/src/components/Header/SearchWrapper.tsx
@@ -1,8 +1,8 @@
 import { Dispatch } from 'react';
 import styled from 'styled-components';
-import Menu from './Menu';
-import MiniSearchBar from './MiniSearchBar';
-import SearchBar from './SearchBar';
+import Menu from 'components/Header/Menu';
+import MiniSearchBar from 'components/Header/SearchBar/MiniSearchBar';
+import SearchBar from 'components/Header/SearchBar/';
 
 type SearchWrapperType = {
   changeSearchBar: (e: React.MouseEvent<HTMLElement>) => void;
